% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges_ci.R
\name{badge_appveyor}
\alias{badge_appveyor}
\title{badge_appveyor}
\usage{
badge_appveyor(repo_full_names)
}
\arguments{
\item{repo_full_names}{vector with combination of username/repo (e.g. 
c("KWB-R/kwb.utils", "KWB-R/kwb.db"))}
}
\value{
appveyor badges for provided repo_full_names
}
\description{
badge_appveyor
}
