% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_codecov.R
\name{badge_codecov}
\alias{badge_codecov}
\title{badge_codecov}
\usage{
badge_codecov(repo_full_names)
}
\arguments{
\item{repo_full_names}{vector with combination of username/repo (e.g. 
c("KWB-R/kwb.utils", "KWB-R/kwb.db"))}
}
\value{
codecov badges for provided repo_full_names
}
\description{
badge_codecov
}
