% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_cran.R
\name{badge_cran}
\alias{badge_cran}
\title{badge_cran}
\usage{
badge_cran(repo_names)
}
\arguments{
\item{repo_names}{vector of repository names (e.g. c("kwb.utils", "kwb.db"))}
}
\value{
crank badges for provided repo_names
}
\description{
badge_cran
}
