% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_dependencies.R
\name{badge_dependencies}
\alias{badge_dependencies}
\title{badge_dependencies}
\usage{
badge_dependencies(repo_names)
}
\arguments{
\item{repo_names}{vector of repository names (e.g. c("kwb.utils", "kwb.db"))}
}
\value{
dependency badges for provided repo_names
}
\description{
badge_dependencies
}
