% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_opencpu.R
\name{badge_opencpu}
\alias{badge_opencpu}
\title{badge_opencpu}
\usage{
badge_opencpu(
  url,
  logo_path = compose_url_githubusercontent(subdomain = "avatars2", path = "u/28672890",
    parameters = list(s = 200, v = 4)),
  size = 24
)
}
\arguments{
\item{url}{url to repository on Gitlab}

\item{logo_path}{path to OpenCpu logo (default: 
"https://avatars2.githubusercontent.com/u/28672890?s=200&v=4")}

\item{size}{size of logo in pixels (default: 24)}
}
\value{
OpenCpu logo in html with path to R package on OpenCpu
}
\description{
badge_opencpu
}
