% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges_ci.R
\name{badge_travis}
\alias{badge_travis}
\title{badge_travis}
\usage{
badge_travis(repo_full_names)
}
\arguments{
\item{repo_full_names}{vector with combination of username/repo (e.g. 
c("KWB-R/kwb.utils", "KWB-R/kwb.db"))}
}
\value{
travis badges for provided repo_full_names
}
\description{
badge_travis
}
