% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_zenodo.R
\name{badge_zenodo}
\alias{badge_zenodo}
\title{badge_zenodo}
\usage{
badge_zenodo(repo_full_names, zenodo_token = get_token("ZENODO"))
}
\arguments{
\item{repo_full_names}{vector with combination of username/repo (e.g.
c("KWB-R/kwb.utils", "KWB-R/kwb.db"))}

\item{zenodo_token}{zenodo authentication token. 
Default: kwb.pkgstatus:::get_token("ZENODO")}
}
\value{
zenodo badges for provided repo_full_names
}
\description{
badge_zenodo
}
