% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_documentation.R
\name{check_docu_dev}
\alias{check_docu_dev}
\title{Check documentation: development}
\usage{
check_docu_dev(repo_names, url = "http://kwb-r.github.io")
}
\arguments{
\item{repo_names}{vector of repository names to be checked}

\item{url}{main url for Github pages (default: "http://kwb-r.github.io")}
}
\value{
character vector with links in case documentation of development 
version for R packages is available
}
\description{
Check documentation: development
}
