% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_documentation.R
\name{check_docu_release}
\alias{check_docu_release}
\title{Check documentation: release}
\usage{
check_docu_release(repo_names, url = "http://kwb-r.github.io")
}
\arguments{
\item{repo_names}{vector of repository names to be checked}

\item{url}{main url for Github pages (default: "http://kwb-r.github.io")}
}
\value{
character vector with links in case documentation of latest release
for R packages is available
}
\description{
Check documentation: release
}
