% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gitlab_backup.R
\name{check_gitlab_backup}
\alias{check_gitlab_backup}
\title{check_gitlab_backup}
\usage{
check_gitlab_backup(
  group = "KWB-R",
  github_token = get_github_token(),
  gitlab_token = get_gitlab_token()
)
}
\arguments{
\item{group}{username or organisation for Github/Gitlab (default: "KWB-R")}

\item{github_token}{github access token (default: Sys.getenv("GITHUB_TOKEN"))}

\item{gitlab_token}{gitlab access token (default: Sys.getenv("GITLAB_TOKEN")))}
}
\value{
data.frame containing all Github repositoriers that are mirrored in 
Gitlab (i.e. were at least syncronised within the last 2 hours)
}
\description{
check_gitlab_backup
}
