% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_opencpu_deploy.R
\name{check_opencpu_deploy}
\alias{check_opencpu_deploy}
\title{Get all Github repos that are deployed on OpenCpu}
\usage{
check_opencpu_deploy(group = "KWB-R")
}
\arguments{
\item{group}{username or organisation for Github (default: "KWB-R")}
}
\value{
data.frame containing names with OpenCpu badges for all Github 
repositories that are deployed on OpenCpu (default: https://kwb-r.ocpu.io)
}
\description{
Direct deployment of R packages (including vignette build) by 
using webhooks as described in OpenCpu blog post 
(https://www.opencpu.org/posts/opencpu-release-1-4-5/) and online help 
(https://www.opencpu.org/cloud.html)
}
