% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_rpackages.R
\name{create_report_rpackages}
\alias{create_report_rpackages}
\title{Create R packages status report}
\usage{
create_report_rpackages(
  secrets_csv = NULL,
  non_r_packages = get_non_r_packages(),
  export_dir = ".",
  input_rmd = system.file("extdata/reports/status_report.Rmd", package = "kwb.pkgstatus")
)
}
\arguments{
\item{secrets_csv}{path to "secrets.csv" file, if "NULL" Sys.env variables 
for the following services are used/need to be defined: APPVEYOR_TOKEN, 
GITHUB_TOKEN, GITLAB_TOKEN, CODECOV_TOKEN, ZENODO_TOKEN, 
(default: NULL)}

\item{non_r_packages}{a character vector with repositories in KWB-R group that
are not R packages (default: \code{get_non_r_packages})}

\item{export_dir}{report export directory (default: ".")}

\item{input_rmd}{default: system.file("extdata/reports/status_report.Rmd", 
package = "kwb.pkgstatus")}
}
\value{
creates html status report for R packages and returns the absolute 
path to the export directory
}
\description{
Create R packages status report
}
