% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverages.R
\name{get_coverage}
\alias{get_coverage}
\title{get_coverage}
\usage{
get_coverage(repo_full_name, codecov_token = get_token("CODECOV"), dbg = TRUE)
}
\arguments{
\item{repo_full_name}{one combination of username/repo (e.g."KWB-R/kwb.db")}

\item{codecov_token}{codecov authentication token.
Default: kwb.pkgstatus:::get_token("CODECOV")}

\item{dbg}{debug if TRUE (default: TRUE)}
}
\value{
codecov coverage in percent for provided repo_full_name
}
\description{
get_coverage
}
