% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_github_package_db.R
\name{make_github_package_db}
\alias{make_github_package_db}
\title{Make Package Database for Packages on GitHub}
\usage{
make_github_package_db(group = "kwb-r", dbg = TRUE)
}
\arguments{
\item{group}{Name of GitHub organisation, defaults to "kwb-r"}

\item{dbg}{logical indicating whether or not to show debug messages. 
The default is \code{TRUE}}
}
\value{
data frame with each column representing a field in the `DESCRIPTION`
  file
}
\description{
ready to be used in e.g. \code{\link[tools]{package_dependencies}}
}
