% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_collections.R
\name{process_hitter_response}
\alias{process_hitter_response}
\title{Helper function for Zenodo}
\usage{
process_hitter_response(response)
}
\arguments{
\item{response}{response provided by httr::content()}
}
\value{
a tibble of available Zenodo data
}
\description{
Helper function for Zenodo
}
