% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{url_success}
\alias{url_success}
\title{url_success}
\usage{
url_success(url)
}
\arguments{
\item{url}{url of documentation website}
}
\value{
TRUE in case HTTP status code is 200, if not: FALSE
}
\description{
url_success
}
