% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_gitlab.R
\name{badge_gitlab}
\alias{badge_gitlab}
\title{badge_gitlab}
\usage{
badge_gitlab(
  url,
  logo_path = paste0("https://gitlab.com/gitlab-com/gitlab-artwork/raw/",
    "master/logo/logo-square.png"),
  size = 24
)
}
\arguments{
\item{url}{url to repository on Gitlab}

\item{logo_path}{path to Gitlab logo (default: 
"https://gitlab.com/gitlab-com/gitlab-artwork/raw/master/logo/logo-square.png")}

\item{size}{size of logo in pixels (default: 24)}
}
\value{
Gitlab logo in html with path to repository in Gitlab
}
\description{
badge_gitlab
}
