% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_license.R
\name{badge_license}
\alias{badge_license}
\title{badge_license}
\usage{
badge_license(license_keys, github_token = get_github_token())
}
\arguments{
\item{license_keys}{one or many valid license keys from c("agpl-3.0", 
"apache-2.0", "bsd-2-clause", "bsd-3-clause", "epl-2.0", "gpl-2.0", "gpl-3.0",    
"lgpl-2.1", "lgpl-3.0", "mit", "mpl-2.0", "unlicense")}

\item{github_token}{github access token. Default: 
kwb.pkgstatus:::get_github_token()}
}
\value{
badge for all provided license keys
}
\description{
badge_license
}
