% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report_rpackages.R
\name{create_report_rpackages}
\alias{create_report_rpackages}
\title{Create R packages status report#'}
\usage{
create_report_rpackages(secrets_csv, non_r_packages = c("kwb-r.github.io",
  "fakin.blog", "fakin.doc", "HydroServerLite", "GeoSalz"), export_dir = ".",
  input_rmd = system.file("extdata/reports/status_report.Rmd", package =
  "kwb.pkgstatus"))
}
\arguments{
\item{secrets_csv}{path to "secrets.csv" file}

\item{non_r_packages}{a character vector with repositories in KWB-R group that
are not R packages c("kwb-r.github.io", "fakin.blog", "fakin.doc", 
"HydroServerLite", "GeoSalz")}

\item{export_dir}{report export directory (default: ".")}

\item{input_rmd}{default: system.file("extdata/reports/status_report.Rmd", 
package = "kwb.pkgstatus")}
}
\value{
creates html status report for R packages
}
\description{
Create R packages status report#'
}
