% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverages.R
\name{get_coverage}
\alias{get_coverage}
\title{get_coverage}
\usage{
get_coverage(
  repo_full_name,
  codecov_token = Sys.getenv("CODECOV_TOKEN"),
  dbg = TRUE
)
}
\arguments{
\item{repo_full_name}{one combination of username/repo (e.g."KWB-R/kwb.db")}

\item{codecov_token}{codecov authentication token (default: 
Sys.getenv("CODECOV_TOKEN"))}

\item{dbg}{debug if TRUE (default: TRUE)}
}
\value{
codecov coverage in percent for provided repo_full_name
}
\description{
get_coverage
}
