% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverages.R
\name{get_coverages}
\alias{get_coverages}
\title{get_coverage}
\usage{
get_coverages(
  repo_full_names,
  codecov_token = Sys.getenv("CODECOV_TOKEN"),
  dbg = TRUE
)
}
\arguments{
\item{repo_full_names}{vector with combination of username/repo (e.g.
c("KWB-R/kwb.utils", "KWB-R/kwb.db"))}

\item{codecov_token}{zenodo authentication token (default: 
Sys.getenv("CODECOV_TOKEN")}

\item{dbg}{debug if TRUE (default: TRUE)}
}
\value{
data.frame with coverage percent and url for all provided 
repo_full_names
}
\description{
get_coverage
}
