% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_infos.R
\name{get_github_repos}
\alias{get_github_repos}
\title{get_github_repos}
\usage{
get_github_repos(group = "KWB-R", github_token = get_github_token())
}
\arguments{
\item{group}{username or organisation for Github (default: "KWB-R")}

\item{github_token}{github access token.
Default: kwb.pkgstatus:::get_github_token()}
}
\value{
data.frame with for all repositories of the user/organisation defined 
in parameter group (private repos will only be accessible if the token is 
configured to allow that)
}
\description{
get_github_repos
}
