% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_infos.R
\name{get_gitlab_repos}
\alias{get_gitlab_repos}
\title{get_gitlab_repos}
\usage{
get_gitlab_repos(group = "KWB-R", gitlab_token = get_gitlab_token())
}
\arguments{
\item{group}{username or organisation for Gitlab (default: "KWB-R")}

\item{gitlab_token}{gitlab access token.
Default: kwb.pkgstatus:::get_gitlab_token()}
}
\value{
data.frame with for all repositories of the user/organisation defined 
in parameter group (private repos will only be accessible if the token is 
configured to allow that)
}
\description{
get_gitlab_repos
}
