% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_of_r_packages_on_github.R
\name{get_names_of_r_packages_on_github}
\alias{get_names_of_r_packages_on_github}
\title{Get Names of R Packages on GitHub}
\usage{
get_names_of_r_packages_on_github(
  group = "kwb-r",
  public = TRUE,
  private = TRUE
)
}
\arguments{
\item{group}{name of organisation on GitHub, defaults to "kwb-r"}

\item{public}{logical indicating whether private repositories are to be
considered. Default: \code{TRUE}}

\item{private}{logical indicating whether public repositories are to be
considered. Default: \code{TRUE}}
}
\value{
vector of character representing the names of the \code{public} 
  and/or \code{private} repositories (as requested), owned by the 
  organisation \code{groupt} on GitHub
}
\description{
Get Names of R Packages on GitHub
}
