% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_r_packages.R
\name{get_non_r_packages}
\alias{get_non_r_packages}
\title{Helper function: get_non_r_packages}
\usage{
get_non_r_packages()
}
\value{
returns vector with KWB-R repos on Github, which are not R packages
}
\description{
Helper function: get_non_r_packages
}
\examples{
get_non_r_packages()

}
