% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_status_rpackages.R
\name{prepare_status_rpackages}
\alias{prepare_status_rpackages}
\title{prepare_status_rpackages}
\usage{
prepare_status_rpackages(secrets_csv, non_r_packages = c("kwb-r.github.io",
  "fakin.blog", "fakin.doc", "HydroServerLite"))
}
\arguments{
\item{secrets_csv}{path to "secrets.csv" file}

\item{non_r_packages}{a character vector with repositories in KWB-R group that
are not R packages c("kwb-r.github.io", "fakin.blog", "fakin.doc", 
"HydroServerLite")}
}
\value{
data.frame with R package status information
}
\description{
prepare_status_rpackages
}
