% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_status_rpackages.R
\name{prepare_status_rpackages}
\alias{prepare_status_rpackages}
\title{prepare_status_rpackages}
\usage{
prepare_status_rpackages(
  secrets_csv = NULL,
  non_r_packages = get_non_r_packages()
)
}
\arguments{
\item{secrets_csv}{path to "secrets.csv" file, if "NULL" Sys.env variables 
for the following services are used/need to be defined: APPVEYOR_TOKEN, 
GITHUB_TOKEN, GITLAB_TOKEN, CODECOV_TOKEN, ZENODO_TOKEN, 
(default: NULL)}

\item{non_r_packages}{a character vector with repositories in KWB-R group that
are not R packages (default: \code{get_non_r_packages})}
}
\value{
data.frame with R package status information
}
\description{
prepare_status_rpackages
}
