% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_zenodo.R
\name{zen_collections}
\alias{zen_collections}
\title{Zenodo: get available collections}
\usage{
zen_collections(access_token = getOption("zenodo_token"))
}
\arguments{
\item{access_token}{= getOption("zenodo_token")}
}
\value{
a tibble of available Zenodo data
}
\description{
Zenodo: get available collections
}
